requirements.txt
# --- requirements.txt ---
Django>=4.2
djangorestframework>=3.14
pillow>=10.0
pytesseract>=0.3.10
celery>=5.3
redis>=5.0
python-dotenv>=1.0
openai>=1.0
anthropic>=0.18
google-generativeai>=0.3
groq>=0.4
drf-yasg>=1.21
django-cors-headers>=4.3
